Make generate-python run from any directory
Previously this script assumed that the current directory was the root
of the codegen tree - which is undesirable if the script is being invoked
as part of some other build.

Also the script used to write to a temporary classpath.txt file, which is
undesirable.  Since Java6 Java has supported looking for all jars in
a particular directory.