Allow SPNEGO security to be enforced as system level
By introducing the Java system property -DapiInvokerAuthScheme=SPNEGO
it is now possible to make sure that security is enforced regardless
of the client settings provided by the client.

This is sometimes needed as the Security Officer requires policies to
be enforced regardless of what the client code is actually going to
provide as parameters.