csharp: Fixed handling of 2xx result codes != 200.
Before this change, API calls returning successful result codes other than 200 were treated as failures. For example, one possible successful result code for a POST or PUT operation is 201 Created; this previously resulted in an exception but is now treated as a success.

A code is classified as a success in a similar way to how javax.ws.rs.core.Response (as of jersey-core 1.8.3) determines a result code's "Family".