Adds default settings for the generated Swagger ruby client
I was having a bit of pain when first using the generated client, though
I'm not entirely sure if this is the correct approach. I just noticed that
when trying to make API requests I needed to call `Swagger.configure` before
things would work, but then it would attempt to call the pets API instead of
my API that I had just conusmed to generate the client.