Unreachable Code in Deserialize
Removing:

if (type.GetType() == typeof(Object))
        return (Object)content;

Because it is unreachable. When you call .GetType() on an instance of Type, which is being done here, you actually are getting the type of the type, which in this case would be RuntimeType. You don't get the type that the Type object is wrapping. Because you always get the type of RuntimeType, this if statement always evaluates to false.

I'm proposing to remove these lines, because it seems it always evaluates to false, and therefore isn't hit. Their original purpose seems undocumented as well, and I wonder if it's some kind of workaround.

A way to have this functionality and have it evaluate properly would be:

if (type.FullName == typeof(Object).FullName)
        return (Object)content;