spring-mvc-j8: Uses Java 8 interface default and provide async callback
This template is mainly for Maven code-gen plugin use-case. When Swagger
spec on existing project changed there is no need to manually copy/paste
the new functions from the generated client. This will provide a default
(empty) implementation to existing impl and user just need to override
it.

Because it is an interface, an actual implementation will be needed to
actuate a service end-point.