[generator] Exclude api/model tests & docs via options
Adds support for system properties apiTests, modelTests, modelTests, modelDocs.
All accepting a boolean value to explicitly define whether or not these
should be generated.

These properties aren't considered "features", so specifying
-DmodelTests=false for example won't cause api or supportFiles to be ignored.

Includes additionalProperty excludeTests for when apiTests and modelTests are
both set to false.

Also includes update to csharp client generator to prevent generation of
the Test project or inclusion of the Test project when both api and
model tests are excluded.

see #2506