[PHP] Add path without expanded path parameters to callApi
In continuation of #3117 it could be useful to know the path of an
endpoint (without path parameters expanded) in the `callApi` method of
`ApiClient`.

This is for use cases where you would create a derived class from
`ApiClient` for manipulating responses from the server before further
processing (#3117) or add extended logging of the API calls.