Add missing spring-cloud client to Swagger Editor
Swagger editor reads the available clients from the backend. Code generators are are categorized as either Server or Client. The spring code generator is categorized as Server therefore it won't be listed as a client. I created a new code generator only so it can be registered as Client. This is clearly a workaround, but a proper solution would require a bigger refactor and I am not sure it worth the effort. If someone does the full refactor then I think a getHumanReadableName() method should be added to CodeGenConfig.java so generators can provide their human readable name. Currently it is calculated by the Swagger editor in
```
$scope.capitalizeGeneratorName
```
which means adding a new language requires change in two different projects