Fix compilation error in Go templates
When all of the pathParams for an API call are empty then the imported
"strings" and "fmt" are unused. This is a compilation error in Go and
the best way to work around it considering the limitations of Mustache
is to add dummy usages of the libraries in the file.