Handle enums when they are declared as definitions.
These appear at the top level of the template (as opposed to as a property
of another class).  Handle these with an enum declaration and a matching
decoder.

Note that decode failure is treated as a fatalError, which I'm not a fan of,
but which matches the rest of the Decoder interface.

Note also that this puts the enum value in double-quotes, assuming that it
is a string on the wire.  I'm pretty sure that this isn't going to work
for cases where enums are declared as some other type, but I don't know how
to fix this yet.