[dart] updated intl library version
The version of the "intl" library is now more than a year old and is starting to conflict with other libraries' dependencies, in our case with angular 2 for dart. Only date parsing, i.e. a very limited subset of "intl" features, is used in the generated libraries. We find it unlikely, that this feature would be changed in a major way in future, and therefore and to avoid this problem in future, we find it justified to set the version constraint to "any".