Allow multiple requests in parallel in Python client
If you tried to do two parallel calls to the same API object in the
Python client you would get an error from urllib3 connection pool:

  Connection pool is full, discarding connection: ***

Because the default maxsize=1:
https://github.com/shazow/urllib3/blob/f9409436f83aeb79fbaf090181cd81b784f1b8ce/urllib3/connectionpool.py#L162

By defaulting to a higher maxsize we mitigate for the common use case
where a user is running a couple of requests in parallel.

Ideally, in the future, this should be a configuration paramater
together with the pool size.