Add hasDiscriminator property to CodegenProperty and CodegenParameter.
CodegenModel's discriminator  can be used to find out if there is a discriminator in a Model.
CodegenOperation's discriminator can be used to find out if returntype contains a discriminator.

CodegenProperty and CodegenParameter does not have discriminator fields to identify if type associated with them contains a discriminator.
This change adds hasDiscriminator fields to CodegenProperty and CodegenParameter which can be used to identify if type associated with them contains a discriminator.