Update html template and handle "default" reponse code (issue #4982)
Add isDefaultCode for CodegenResponse to indicate iff the response is "default"
(The existing isDefault is true for that case *or* if the response is the first 2xx response.)

Update index.mustache to emit "default" instead of 0 if the response key is "default".
Add the model title to the model TOC and model section heading.