[JAVA][jaxrs-spec] Add support for generating type-safe query parameters (#5075)
All the logic is already in place so I just make sure parameters in:query
also insert an enum in the generated API and that the type for the REST
operation parameter is that enum (and not just String as it was before).

The code is somewhat ugly and its functionality is still limited: for example
Container<Container<String>> param; (or member)
won't work. I think it's enough for now to use this for the regular cases
when only List<Integer> and List<String> are being passed around.

The state of the if-variable-then-variable if-notVariable-then-other
is quite ludicrous now in enumClass.mustache and queryParams.mustache.
It's probably a good idea to refactor that.