[csharp] Fix $ref model in parameter as additionalProperties
Possible fix for 5132 in which a RefModel defined as
additionalProperties on a body parameter was being ignored.

I'm not entirely sure this is an appropriate fix because this assumes
additionalProperties.$ref refers to *the* model definition, rather than
a mixin of properties to be applied to some dictionary of key/values.