#6718: Add exclusive min and max values for DecimalMin annotation
Swagger has the ability to express whether a minimum or maximum value is inclusive or exclusive (via exclusiveMinimum and exclusiveMaximum). The @DecimalMin and @DecimalMax annotations are inclusive by default (`inclusive = true`), but can express exclusive limits with the parameter `inclusive = false`. This maps Swagger's exclusiveMinimum to the `inclusive = false` parameter.