fix: unify response mapping (dehydrating)
This adds a more unified handling for response code to type mapping.
This fixes the deserialization for non default responses, which in
previously implementation were only based on exception logic.
This approach assumes that when the specification defines an possible
response code outside of the default one, the client should handle the
result in the same way as the default, but still propagate the result
as exception. This exception is a MappedException, as it's defined by
specification. Any other unexpected response is still an ApiException.

It's a BC break.
- The model if ApiException is changed, as it's now more
generic.
- The client makes use of php 7.1 features, I assume it's a required
  step. (https://secure.php.net/supported-versions.php)
  This includes return type hinting. (For operations with 'void' only.)