Fix issue with date formatting
If the user is in a different region, then their default NSDateFormatter will use their device's region. Since we are sending back a standard date object from the US, we need to let the formatter know where the date is coming from. Once we set the locale we are confirming that the server will only send back US dates and thus the phone will be able to parse them appropriately.