[cpprest] Make the generated APIs mockable.
We want to able to unit test our code that uses the generated API code.
In order to do that we need a virtual interface to mock away the
generated API code.

This PR introduces two new config options for the cpprest generator:
* "generateInterfacesForApis" will generate an abstract base class
  (interface) for all APIs.
* "generateGMocksForApis" will additionally generate Google Mock classes
  for the APIs. This config option of course implies the first one.

If the options are not set only minor white space changes will result
(see the changes in the petstore example).