Add `CodegenConfig.shouldSkipModel()` method
The addition of this method allows a generator to specify that specific
models should not generate an output file. This also replaces the
hardcoded check in `DefaultGenerator` for skipping alias models if the
language is `java` or `inflector`. There are many more Java-based
'languages` which need this functionality - it should be possible for
the generators to control skipped models themselves.

Note that the skipping of alias models in Java-based languages should
now be controlled downstream in `swagger-codegen-generators`.