[csharp] Fix string enum models sometimes losing attributes
Depending on HashMap iteration order, the isString flag added to string
enum values can be lost due to copying a reference to the
allowableValues map to properties that reference the enum. The result is
that the code fails to compile.

This should fix #7656, fix #7657 (regarded as a duplicate of 7656), and
fix #8027 (which also appears to be the same issue).