Add nullable property support from OAS 3.0 spec
When I wrote this, I missed the changes on CodegenParameter and I had
some unnecessary changes on CodegenModel. I realized my mistake when I
went to add a similar change to openapi-generator and saw it was already
there, and they knew more about what to do than I did.

So, this is based in part on work by @wing328 and @jmini
- OpenAPITools/openapi-generator#873
- OpenAPITools/openapi-generator#889
- OpenAPITools/openapi-generator#930