README.md: added minor details to Docker section
The docker section did not currently contain the repositories for v3 nor any indication that listed repositories were for v2. I only found this out thanks to #8690, so I thought it might help others to have that listed.

I also fixed a little link in the TOC that was not properly linking to the section below.