Adding csproj file to csharp-dotnet2 code generation
The `csproj` file is a more convenient way for building and publishing
packages in any platform by simply calling `dotnet build` command.

This file also allows publishing to a NuGet Package Server without
requiring the interaction with a `nuspec` file, making it easy for users
of the code generation tool to build and publish any generated package.

With this change, the Mono installation script becomes deprecated,
although it will be kept for backward-compatibility purposes.

This closes #8334