[Dart] Add sdk constraints to pubspec template
SDK constraints are required for Dart 2 projects, and optional for Dart
1. The actual constraint values match the dependency on http.

http 0.11.1 has a minimum sdk version of 1.1.0, so we have matched this
constraint. A maximum sdk version has been set at 3.0.0, as this is
standard practice for Dart 2 projects, as Dart 3 has not yet been
released and is therefore untested.

Closes #9422