[issue-9638] fix issue-9638
Use the properties from the referenced Model when parsing variables of a
model.  This way detailed type information will be passed to the
templates (isNumeric, isDate, etc).

Remove the now unnecessary fix for issue-7980 which solved only part of
the underlying issue.

The original fix for issue-7980 was not consistent with inline
properties, updated the test case to reflect this.