[PYTHON] use isoparse instead of parse for datetime fields
Currently, the python-generated client will deserialize a datetime field
by using dateutil.parser.parse. This is a less-strict parser, allowing
it to support more than just ISO-8601 formats, but also has some weird
behavior of setting the timezone to tzlocal instead of tzutc on machines
that have utc as their timezone (only occasionally though).

dateutil.parser.isoparse is a faster and stricter ISO-8601 parser, and
the swagger spec for date-time fields indicates that it should follow
ISO-8601. Therefore, to remove the possibility of tzutc being
interpreted as tzlocal (which can be confusing and unexpected), we
should use isoparse.

This change modifies the mustache template for python code-gen to use
isoparse instead of parse. The result is the generated python code will
correctly report timezones as tzutc on machines that used to report
tzlocal. It also runs almost an order of magnitude faster!