Add okhttp4-gson codegen
Add support for okhttp4-gson as a replacement for okhttp-gson

okhttp-gson is based on okhttp 2.x which is no longer supported. okhttp 3.x is also [not supported anymore](https://square.github.io/okhttp/security/security/)

Creating a new target for codegen since okhttp 4.x is a breaking change over the okhttp 2.x with the changed package names.

The overall change is similar to the existing code with a few updates
- updates due to deprecations documented in https://square.github.io/okhttp/changelogs/changelog_3x/#version-300-rc1
- kept the utility methods in the existing code as-is

tested `mvn package`, ran the petstore generation scripts and the integration tests that are generated

Fixes #9606