Fix issue #12071.
Adjust the DefaultGenerator class so that the private method
generateApis() runs partialy regardless of whether or not we are
generating APIs. The part that always runs computes API data that is
used by some supporting files templates.

This affects the the Go client target among others.

Allow supporting files that refer to the API to be generated properly.