[python] - update python-petstore security
Run ./bin/security/python-petstore.sh to pick up basic auth fixes.

Note that there have been other template changes in python but
security/python-petstore.sh was not run so the sample is out of date. I
have intentionally excluded the changes that are not related to the
basic auth fix.

Signed-off-by: Cory Todd <cory.todd@canonical.com>